requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
asyncpg==0.29.0
openai==1.70.0
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.25.2
starlette==0.27.0


